feat(gui): move filter to footer and actions from footer to menu bar
