+++ api-editor/gui/src/features/menuBar/MenuBar.tsx
@@ -12,20 +12,28 @@ import {
     MenuItemOption,
     MenuList,
     MenuOptionGroup,
-    Spacer,
     useColorMode,
 } from '@chakra-ui/react';
 import React from 'react';
 import { FaChevronDown } from 'react-icons/fa';
 import { useAppDispatch, useAppSelector } from '../../app/hooks';
-import { selectAnnotationStore, selectUsernameIsValid } from '../annotations/annotationSlice';
+import {
+    AnnotationStore,
+    redo,
+    selectAnnotationStore,
+    selectUsernameIsValid,
+    undo,
+} from '../annotations/annotationSlice';
 import {
     BatchMode,
     HeatMapMode,
     selectExpandDocumentationByDefault,
+    selectFilter,
     selectHeatMapMode,
     selectShowStatistics,
     selectSortingMode,
+    setAllCollapsedInTreeView,
+    setAllExpandedInTreeView,
     setBatchMode,
     setHeatMapMode,
     setSortingMode,
@@ -38,7 +46,17 @@ import {
 } from '../ui/uiSlice';
 import { DeleteAllAnnotations } from './DeleteAllAnnotations';
 import { GenerateAdapters } from './GenerateAdapters';
-import { FilterControls } from '../filter/FilterControls';
+import { PythonPackage } from '../packageData/model/PythonPackage';
+import { PythonDeclaration } from '../packageData/model/PythonDeclaration';
+import { AbstractPythonFilter } from '../filter/model/AbstractPythonFilter';
+import { UsageCountStore } from '../usages/model/UsageCountStore';
+import {
+    selectFilteredPythonPackage,
+    selectFlatSortedDeclarationList,
+    selectRawPythonPackage,
+} from '../packageData/apiSlice';
+import { selectUsages } from '../usages/usageSlice';
+import { useNavigate, useLocation } from 'react-router-dom';
 
 interface MenuBarProps {
     displayInferErrors: (errors: string[]) => void;
@@ -47,6 +65,7 @@ interface MenuBarProps {
 export const MenuBar: React.FC<MenuBarProps> = function ({ displayInferErrors }) {
     const { colorMode, toggleColorMode } = useColorMode();
     const dispatch = useAppDispatch();
+    const navigate = useNavigate();
 
     const annotationStore = useAppSelector(selectAnnotationStore);
     const sortingMode = useAppSelector(selectSortingMode);
@@ -54,6 +73,14 @@ export const MenuBar: React.FC<MenuBarProps> = function ({ displayInferErrors })
     const showStatistics = useAppSelector(selectShowStatistics);
     const usernameIsValid = useAppSelector(selectUsernameIsValid);
     const expandDocumentationByDefault = useAppSelector(selectExpandDocumentationByDefault);
+
+    const allDeclarations = useAppSelector(selectFlatSortedDeclarationList);
+    const pythonPackage = useAppSelector(selectFilteredPythonPackage);
+    const rawPythonPackage = useAppSelector(selectRawPythonPackage);
+    const pythonFilter = useAppSelector(selectFilter);
+    const annotations = useAppSelector(selectAnnotationStore);
+    const usages = useAppSelector(selectUsages);
+    const declaration = rawPythonPackage.getDeclarationById(useLocation().pathname.split('/').splice(1).join('/'));
 
     const exportAnnotations = () => {
         const a = document.createElement('a');
@@ -229,9 +256,202 @@ export const MenuBar: React.FC<MenuBarProps> = function ({ displayInferErrors })
                 </Box>
             </HStack>
 
-            <Spacer />
+            <Button
+                accessKey="p"
+                disabled={!declaration}
+                onClick={() => {
+                    let navStr = getPreviousElementPath(
+                        allDeclarations,
+                        declaration!,
+                        pythonFilter,
+                        annotations,
+                        usages,
+                    );
+                    if (navStr !== null) {
+                        //navigate to element
+                        navigate(`/${navStr}`);
 
-            <FilterControls />
+                        //update tree selection
+                        const parents = getAncestors(navStr, pythonPackage);
+                        dispatch(setAllExpandedInTreeView(parents));
+                    }
+                }}
+            >
+                Previous
+            </Button>
+            <Button
+                accessKey="n"
+                disabled={!declaration}
+                onClick={() => {
+                    let navStr = getNextElementPath(
+                        allDeclarations,
+                        declaration!,
+                        pythonFilter,
+                        annotations,
+                        usages,
+                    );
+                    if (navStr !== null) {
+                        //navigate to element
+                        navigate(`/${navStr}`);
+
+                        //update tree selection
+                        const parents = getAncestors(navStr, pythonPackage);
+                        dispatch(setAllExpandedInTreeView(parents));
+                    }
+                }}
+            >
+                Next
+            </Button>
+
+            <Button
+                accessKey="u"
+                onClick={() => {
+                    const parent = declaration?.parent();
+                    if (parent && !(parent instanceof PythonPackage)) {
+                        navigate(`/${parent.id}`);
+                    }
+                }}
+            >
+                Go to Parent
+            </Button>
+
+            <Button
+                onClick={() => {
+                    dispatch(setAllExpandedInTreeView(getDescendantsOrSelf(pythonPackage)));
+                }}
+            >
+                Expand All
+            </Button>
+            <Button
+                onClick={() => {
+                    dispatch(setAllCollapsedInTreeView(getDescendantsOrSelf(pythonPackage)));
+                }}
+            >
+                Collapse All
+            </Button>
+
+            <Button
+                disabled={!declaration}
+                onClick={() => {
+                    dispatch(setAllExpandedInTreeView(getDescendantsOrSelf(declaration!)));
+                }}
+            >
+                Expand Selected
+            </Button>
+            <Button
+                disabled={!declaration}
+                onClick={() => {
+                    dispatch(setAllCollapsedInTreeView(getDescendantsOrSelf(declaration!)));
+                }}
+            >
+                Collapse Selected
+            </Button>
+            <Button
+                accessKey="z"
+                disabled={!declaration}
+                onClick={() => {
+                    dispatch(undo());
+                }}
+            >
+                Undo
+            </Button>
+            <Button
+                accessKey="y"
+                disabled={!declaration}
+                onClick={() => {
+                    dispatch(redo());
+                }}
+            >
+                Redo
+            </Button>
         </Flex>
     );
+};
+
+const getPreviousElementPath = function (
+    declarations: PythonDeclaration[],
+    start: PythonDeclaration,
+    filter: AbstractPythonFilter,
+    annotations: AnnotationStore,
+    usages: UsageCountStore,
+): string {
+    let currentIndex = getPreviousIndex(declarations, getIndex(declarations, start));
+    let current = getElementAtIndex(declarations, currentIndex);
+    while (current !== null && current !== start) {
+        if (filter.shouldKeepDeclaration(current, annotations, usages)) {
+            return current.id;
+        }
+        currentIndex = getPreviousIndex(declarations, currentIndex);
+        current = getElementAtIndex(declarations, currentIndex);
+    }
+    return start.id;
+};
+
+const getNextElementPath = function (
+    declarations: PythonDeclaration[],
+    start: PythonDeclaration,
+    filter: AbstractPythonFilter,
+    annotations: AnnotationStore,
+    usages: UsageCountStore,
+): string {
+    let currentIndex = getNextIndex(declarations, getIndex(declarations, start));
+    let current = getElementAtIndex(declarations, currentIndex);
+    while (current !== null && current !== start) {
+        if (filter.shouldKeepDeclaration(current, annotations, usages)) {
+            return current.id;
+        }
+        currentIndex = getNextIndex(declarations, currentIndex);
+        current = getElementAtIndex(declarations, currentIndex);
+    }
+    return start.id;
+};
+
+const getPreviousIndex = function (declarations: PythonDeclaration[], currentIndex: number | null): number | null {
+    if (currentIndex === null || currentIndex < 0 || currentIndex >= declarations.length) {
+        return null;
+    }
+
+    return (currentIndex - 1 + declarations.length) % declarations.length;
+};
+
+const getNextIndex = function (declarations: PythonDeclaration[], currentIndex: number | null): number | null {
+    if (currentIndex === null || currentIndex < 0 || currentIndex >= declarations.length) {
+        return null;
+    }
+
+    return (currentIndex + 1) % declarations.length;
+};
+
+const getIndex = function (declarations: PythonDeclaration[], current: PythonDeclaration): number | null {
+    const index = declarations.findIndex((it) => it.id === current.id);
+    if (index === -1) {
+        return null;
+    }
+    return index;
+};
+
+const getElementAtIndex = function (declarations: PythonDeclaration[], index: number | null): PythonDeclaration | null {
+    if (index === null) {
+        return null;
+    }
+    return declarations[index];
+};
+
+const getAncestors = function (navStr: string, filteredPythonPackage: PythonPackage): string[] {
+    const ancestors: string[] = [];
+
+    let currentElement = filteredPythonPackage.getDeclarationById(navStr);
+    if (currentElement) {
+        currentElement = currentElement.parent();
+        while (currentElement) {
+            ancestors.push(currentElement.id);
+            currentElement = currentElement.parent();
+        }
+    }
+
+    return ancestors;
+};
+
+const getDescendantsOrSelf = function (current: PythonDeclaration): string[] {
+    return [...current.descendantsOrSelf()].map((descendant) => descendant.id);
 };
